+++ test/unit/credentials-param-test.js
@@ -38,6 +38,7 @@ test('destroy credentials accepted', async t => {
   try {
 
     await destroy({
+      quiet: true,
       appname: 'test-app',
       env: 'staging',
       credentials: {